[UPDATE]create a component directory
