+++ src/app/components/battery-level/battery-level.service.ts
@@ -1,0 +1,60 @@
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
+import { BluetoothCore } from '@manekinekko/angular-web-bluetooth';
+
+@Injectable()
+export class BatteryLevelService {
+  static GATT_CHARACTERISTIC_BATTERY_LEVEL = 'battery_level';
+  static GATT_PRIMARY_SERVICE = 'battery_service';
+
+  constructor(public ble: BluetoothCore) {}
+
+  getFakeValue() {
+    this.ble.fakeNext();
+  }
+
+  getDevice() {
+    return this.ble.getDevice$();
+  }
+
+  streamValues() {
+    return this.ble.streamValues$().map((value: DataView) => value.getUint8(0));
+  }
+
+  /**
+   * Get Battery Level GATT Characteristic value.
+   * This logic is specific to this service, this is why we can't abstract it elsewhere.
+   * The developer is free to provide any service, and characteristics she wants.
+   *
+   * @return {Observable<Number>} Emites the value of the requested service read from the device
+   */
+  getBatteryLevel() {
+    console.log('Getting Battery Service...');
+
+    try {
+      return this.ble
+        .discover$({
+          acceptAllDevices: true,
+          optionalServices: [BatteryLevelService.GATT_PRIMARY_SERVICE]
+        })
+        .mergeMap((gatt: BluetoothRemoteGATTServer) => {
+          return this.ble.getPrimaryService$(
+            gatt,
+            BatteryLevelService.GATT_PRIMARY_SERVICE
+          );
+        })
+        .mergeMap((primaryService: BluetoothRemoteGATTService) => {
+          return this.ble.getCharacteristic$(
+            primaryService,
+            BatteryLevelService.GATT_CHARACTERISTIC_BATTERY_LEVEL
+          );
+        })
+        .mergeMap((characteristic: BluetoothRemoteGATTCharacteristic) => {
+          return this.ble.readValue$(characteristic);
+        })
+        .map((value: DataView) => value.getUint8(0));
+    } catch (e) {
+      console.error('Oops! can not read value from %s');
+    }
+  }
+}
